add, delete, get availability slots

+++ Find_Your_Home/Services/AvailabilitySlotService/AvailabilitySlotService.cs
@@ -1,6 +1,6 @@
-﻿using Find_Your_Home.Models.Bookings;
+﻿using Find_Your_Home.Exceptions;
+using Find_Your_Home.Models.Bookings;
 using Find_Your_Home.Repositories.AvailabilitySlotRepository;
-using Find_Your_Home.Repositories.UserRepository;
 using Find_Your_Home.Services.PropertyService;
 using Find_Your_Home.Services.UserService;
 
@@ -11,38 +11,32 @@ namespace Find_Your_Home.Services.AvailabilitySlotService
         private readonly IUserService _userService;
         private readonly IPropertyService _propertyService;
         private readonly IAvailabilitySlotRepository _availabilitySlotRepository;
-        
+
         public AvailabilitySlotService(IUserService userService, IAvailabilitySlotRepository availabilitySlotRepository, IPropertyService propertyService)
         {
             _userService = userService;
             _availabilitySlotRepository = availabilitySlotRepository;
             _propertyService = propertyService;
         }
-        
+
         public async Task<bool> IsUserOwnerOfProperty(Guid propertyId, Guid userId)
         {
             var user = await _userService.GetUserById(userId);
             if (user == null)
-            {
                 return false;
-            }
 
             var property = await _propertyService.GetPropertyByID(propertyId);
             if (property == null)
-            {
                 return false;
-            }
 
             return property.OwnerId == user.Id;
         }
-        
+
         public async Task<bool> CheckSlotOverlap(Guid propertyId, DateTime date, DateTime startDateTime, DateTime endDateTime)
         {
             var slots = await _availabilitySlotRepository.GetAvailabilitySlotsByPropertyId(propertyId);
             if (slots == null || !slots.Any())
-            {
                 return false;
-            }
 
             foreach (var slot in slots)
             {
@@ -50,9 +44,9 @@ namespace Find_Your_Home.Services.AvailabilitySlotService
                 DateTime slotEnd = slot.Date.Date + slot.EndTime;
 
                 if (slot.Date.Date == date.Date &&
-                    ((startDateTime < slotEnd && endDateTime > slotStart)))
+                    (startDateTime < slotEnd && endDateTime > slotStart))
                 {
-                    return true; 
+                    return true;
                 }
             }
 
@@ -69,9 +63,7 @@ namespace Find_Your_Home.Services.AvailabilitySlotService
         {
             var slots = await _availabilitySlotRepository.GetAvailabilitySlotsByPropertyId(propertyId);
             if (slots == null || !slots.Any())
-            {
-                return null;
-            }
+                throw new AppException("NO_SLOTS_FOUND_FOR_PROPERTY");
 
             return slots.ToList();
         }
@@ -80,25 +72,23 @@ namespace Find_Your_Home.Services.AvailabilitySlotService
         {
             var slot = await _availabilitySlotRepository.GetAvailabilitySlotByIdAsync(availabilitySlotId);
             if (slot == null)
-            {
-                return null;
-            }
+                throw new AppException("SLOT_NOT_FOUND");
 
             return slot;
         }
 
-        public async Task<bool> DeleteAvailabilitySlot(Guid availabilitySlotId)
+        public async Task DeleteAvailabilitySlot(Guid availabilitySlotId)
         {
-            var slot = await _availabilitySlotRepository.GetAvailabilitySlotByIdAsync(availabilitySlotId);
+            var slot = await _availabilitySlotRepository.GetAvailabilitySlotWithBookingsByIdAsync(availabilitySlotId);
+
             if (slot == null)
-            {
-                return false;
-            }
+                throw new AppException("SLOT_NOT_FOUND");
+
+            if (slot.Bookings.Any())
+                throw new AppException("CANNOT_DELETE_SLOT_WITH_BOOKINGS");
 
             _availabilitySlotRepository.Delete(slot);
             await _availabilitySlotRepository.SaveAsync();
-            return true;
         }
-
     }
 }
